Refactor Search components to match other conventions
