+++ react-components/src/data/search/messages.js
@@ -1,30 +1,33 @@
 export default {
-    cancelBtn: 'Cancel',
-    createdWithin: 'Created within',
-    emptyDropDownText: '---',
-    enterCyVerseUserName: 'Enter CyVerse user name',
-    emptyText: 'Enter values...',
-    fileSizeGreater: 'File size is bigger than or equal to',
-    fileSizeLessThan: 'File size is smaller than or equal to',
-    fileSizes: [
-        'KB',
-        'MB',
-        'GB',
-        'TB',
-    ],
-    filterName: 'Filter Name',
-    includeTrash: 'Include items in Trash',
-    metadataAttributeHas: 'Metadata attribute has the words',
-    metadataValueHas: 'Metadata value has the words',
-    modifiedWithin: 'Modified within',
-    nameHas: 'File/Folder name has the words',
-    nameHasNot: "File/Folder name doesn't have",
-    ownedBy: 'Owned by',
-    requiredField: 'This field is required',
-    saveBtn: 'Save',
-    saveSearchBtn: 'Save Search',
-    saveSearchTitle: 'Save Filter',
-    searchBtn: 'Search',
-    sharedWith: 'Shared with',
-    taggedWith: 'Tagged with',
+    locales: 'en-US',
+    messages: {
+        cancelBtn: 'Cancel',
+        createdWithin: 'Created within',
+        emptyDropDownText: '---',
+        enterCyVerseUserName: 'Enter CyVerse user name',
+        emptyText: 'Enter values...',
+        fileSizeGreater: 'File size is bigger than or equal to',
+        fileSizeLessThan: 'File size is smaller than or equal to',
+        fileSizes: [
+            'KB',
+            'MB',
+            'GB',
+            'TB',
+        ],
+        filterName: 'Filter Name',
+        includeTrash: 'Include items in Trash',
+        metadataAttributeHas: 'Metadata attribute has the words',
+        metadataValueHas: 'Metadata value has the words',
+        modifiedWithin: 'Modified within',
+        nameHas: 'File/Folder name has the words',
+        nameHasNot: "File/Folder name doesn't have",
+        ownedBy: 'Owned by',
+        requiredField: 'This field is required',
+        saveBtn: 'Save',
+        saveSearchBtn: 'Save Search',
+        saveSearchTitle: 'Save Filter',
+        searchBtn: 'Search',
+        sharedWith: 'Shared with',
+        taggedWith: 'Tagged with'
+    }
 };
