+++ react-components/src/util/StoreProvider.js
@@ -1,8 +1,8 @@
-import React, {Component} from 'react';
-import {Provider} from "react-redux";
+import React, { Component } from 'react';
+import { Provider } from "react-redux";
 import store from '../store';
 
-export function withStoreProvider(WrappedComponent) {
+export default function withStoreProvider(WrappedComponent) {
     return class extends Component {
         render() {
             return (
